chore: deploy command script
