+++ src/index.js
@@ -26,29 +26,6 @@ const discordClient = new Client({
 discordClient.commands = new Collection();
 discordClient.commands.set(yogalog.data.name, yogalog);
 
-// It is too much for Discord to deploy every time the code is changed
-if (process.env.DEPLOY_COMMANDS === "true") {
-  // Construct and prepare an instance of the REST module
-  const rest = new REST({ version: "10" }).setToken(DISCORD_CLIENT_TOKEN);
-
-  try {
-    const commands = [yogalog.data.toJSON()];
-    console.log(
-      `Started refreshing ${commands.length} application (/) commands.`
-    );
-
-    // The put method is used to fully refresh all commands in the guild with the current set
-    const data = await rest.put(Routes.applicationCommands(DISCORD_CLIENT_ID), {
-      body: commands,
-    });
-
-    console.log(`Successfully reloaded ${data.length} application (/) commands.`);
-  } catch (error) {
-    // And of course, make sure you catch and log any errors!
-    console.error(error);
-  }
-}
-
 // When the client is ready, run this code (only once)
 // We use 'c' for the event parameter to keep it separate from the already defined 'client'
 discordClient.once(Events.ClientReady, (c) => {
